Refactor buildReverbComponentTrackingModel
