+++ src/app/components/PageViewTracking/params/index.ts
@@ -416,45 +416,3 @@ export const getReverbParams = ({
   const params = getATIParams({ atiData, requestContext, serviceContext });
   return buildReverbPageViewModel(params);
 };
-
-export const buildReverbComponentTrackingModel = ({
-  pageIdentifier,
-  producerName,
-  statsDestination,
-  componentName,
-  campaignID,
-  format,
-  type,
-  advertiserID,
-  url,
-}: ComponentTrackingProps) => {
-  const eventDetails = {
-    eventName: type === 'view' ? 'sectionView' : 'sectionClick',
-    eventPublisher: type === 'click' ? 'click' : 'impression',
-    componentName,
-    container: campaignID,
-    attribute: componentName,
-    metadata: format,
-    placement: pageIdentifier,
-    source: advertiserID,
-    result: url,
-    isClick: type === 'click',
-  };
-
-  return {
-    params: {
-      page: {
-        destination: statsDestination,
-        name: pageIdentifier,
-        producer: producerName,
-        additionalProperties: {
-          type: 'AT',
-        },
-      },
-      user: {
-        isSignedIn: false,
-      },
-    },
-    eventDetails,
-  };
-};
