+++ src/app/lib/analyticsUtils/sendComponentTrackingBeacon/index.ts
@@ -1,5 +1,5 @@
-import { buildReverbComponentTrackingModel } from '#app/components/PageViewTracking/params';
 import buildATIComponentTrackingURL from '#app/lib/analyticsUtils/buildATIComponentTrackingURL';
+import buildReverbComponentTrackingModel from '../buildReverbComponentTrackingModel';
 import sendBeacon from '../sendBeacon';
 import { ComponentTrackingProps } from '../types';
